Move String methods (toString, compare, and toUpper/Lower) to StringUtils
