+++ lib/api/ContextFieldsProvider.cpp
@@ -7,7 +7,8 @@
 #include "ContextFieldsProvider.hpp"
 #include "LogSessionData.hpp"
 
-#include "utils/Utils.hpp"
+#include "utils/StringUtils.hpp"
+// #include "utils/Utils.hpp"
 
 namespace MAT_NS_BEGIN
 {
