+++ lib/pal/desktop/WindowsDesktopDeviceInformationImpl.cpp
@@ -8,6 +8,7 @@
 #endif
 #include "pal/PAL.hpp"
 #include "pal/DeviceInformationImpl.hpp"
+#include "utils/StringUtils.hpp"
 #include "WindowsEnvironmentInfo.hpp"
 
 MATSDK_LOG_INST_COMPONENT_NS("DeviceInfo", "Win32 Desktop Device Information")
