+++ lib/stats/MetaStats.cpp
@@ -5,7 +5,7 @@
 
 #include "MetaStats.hpp"
 
-#include <utils/Utils.hpp>
+#include <utils/StringUtils.hpp>
 
 namespace MAT_NS_BEGIN {
